requirements.txt
alembic==1.11.1
blinker==1.6.2
click==8.1.6
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
WTForms==3.0.1
Faker==19.3.1
python-dotenv==1.0.0
python-decouple==3.8
flask-mysql==2.1.3